Update :- See Description ( VampStein )
1. All types of billboard working need to link with sprite Sheets correctly.
2. added in map factory functions.

+++ Engine/Code/Engine/Math/MathUtils.cpp
@@ -1403,7 +1403,7 @@ Vec2 GetSupportPoint( const Vec2* vertices , size_t count , Vec2 direction )
 		if ( product > maxProduct )
 		{
 			maxProduct = product;
-			index = i;
+			index = ( int ) i;
 		}
 	}
 	return vertices[ index ];
